4.21
Delete only own blogs

+++ part4/blog/controllers/blogs.js
@@ -45,7 +45,27 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).end()
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  if(!user) {
+    return response.status(400).json({ error: 'userId missing or not valid' })
+  }
+
+  if(user._id.toString() !== blog.user.toString()) {
+    return response.status(401).json({ error: 'only the creator of the blog can delete it' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
+
   response.status(204).end()
 })
